Fixed namespaces after moving the files.
