+++ MessageCounterFrontend/MainWindowOperations/FileOpener.cs
@@ -1,12 +1,9 @@
 ﻿using MessageCounterBackend;
-using MessageCounterFrontend.InterfaceBackend;
+using MessageCounterFrontend.InterfaceBackend.FileOperators;
+using MessageCounterFrontend.Pages;
 using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 
 namespace MessageCounterFrontend.MainWindowOperations
