+++ MessageCounterFrontend/Pages/StatsPages/OneItemPages/PersonPage.xaml
@@ -1,9 +1,8 @@
-﻿<Page x:Class="MessageCounterFrontend.StatsPages.OneItemPages.PersonPage"
+﻿<Page x:Class="MessageCounterFrontend.Pages.StatsPages.OneItemPages.PersonPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:MessageCounterFrontend.StatsPages.OneItemPages"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="PersonPage">
