+++ MessageCounterFrontend/Pages/StatsPages/PeoplePage.xaml
@@ -1,9 +1,8 @@
-﻿<Page x:Class="MessageCounterFrontend.StatsPage.PeoplePage"
+﻿<Page x:Class="MessageCounterFrontend.Pages.StatsPages.PeoplePage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:MessageCounterFrontend.StatsPage"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       mc:Ignorable="d" 
       d:DesignHeight="400" d:DesignWidth="1000"
       Title="PeoplePage">
