+++ MessageCounterFrontend/Windows/SettingsWindows/SettingsWindow.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 
-namespace MessageCounterFrontend.SettingsWindows
+namespace MessageCounterFrontend.Windows.SettingsWindows
 {
     public abstract class SettingsWindow : Window
     {
