+++ MessageCounterFrontend/Windows/SettingsWindows/WordsSettings.xaml
@@ -1,12 +1,11 @@
 ﻿<base:SettingsWindow
-        xmlns:base="clr-namespace:MessageCounterFrontend.SettingsWindows"
-        x:Class="MessageCounterFrontend.SettingsWindows.WordsSettings"
+        xmlns:base="clr-namespace:MessageCounterFrontend.Windows.SettingsWindows"
+        x:Class="MessageCounterFrontend.Windows.SettingsWindows.WordsSettings"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MessageCounterFrontend.SettingsWindows"
-        mc:Ignorable="d" 
+        mc:Ignorable="d"
         Height="250" Width="400" 
         ResizeMode="NoResize" WindowStyle="None"
         WindowStartupLocation="CenterOwner"
